refactor: RNDEAS-964: update to use new image interface
Update plugins which use `load_image` to work with the new image interface.

The use of ImageAdaptor means the plugins would still work as is, but only if `image_converters` apconfig is configured to not return a WandImage instance.

This update the plugins to work with any of the converters

Also updates the extended image plugin to use `load_image`, rather than extending `PillowThumbnail`, which explicitly uses Pillow.